+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Cleaned up the auth sign up validation schema
+
 ---
 
 ## [1.0.2] -2025-09-07